+++ CHANGELOG.rst
@@ -45,6 +45,8 @@ Fixed
 - Use default values of ``replication_connect_quorum`` and
   ``replication_sync_timeout`` in first ``box.cfg`` call.
 
+- Check hash of cluster cookie on stateful failover configuration
+  (`#1765 <https://github.com/tarantool/cartridge/issues/1765>`_).
 -------------------------------------------------------------------------------
 [2.7.7] - 2022-12-09
 -------------------------------------------------------------------------------